pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devscontext"
version = "0.1.0"
description = "MCP server that provides AI coding agents with synthesized engineering context from SDLC tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "you@devscontext.com" },
]
keywords = ["mcp", "ai", "context", "jira", "claude-code", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]

dependencies = [
    "mcp>=1.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "click>=8.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
all = ["anthropic>=0.40", "openai>=1.50"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
    "ruff>=0.8",
    "mypy>=1.10",
]

[project.scripts]
devscontext = "devscontext.cli:main"

[project.urls]
Homepage = "https://devscontext.com"
Repository = "https://github.com/yourusername/devscontext"
Documentation = "https://devscontext.com/docs"

[tool.hatch.build.targets.wheel]
packages = ["src/devscontext"]

# =============================================================================
# RUFF CONFIGURATION
# =============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # ruff-specific rules
]
ignore = [
    "ANN401",  # Dynamically typed expressions (Any) are disallowed
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "S101"]  # No annotations required in tests, allow assert

[tool.ruff.lint.isort]
known-first-party = ["devscontext"]

# =============================================================================
# MYPY CONFIGURATION
# =============================================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "mcp.*",
    "yaml",
]
ignore_missing_imports = true

# =============================================================================
# PYTEST CONFIGURATION
# =============================================================================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
